Implementation of complete GPU Proximity Engine
- LBVH based tet level broad phase
- Narrow phase is now complicated indexing free and just takes array of
  pairs of tets
- Benchmark plotting scripts
- Extensive timing code for CPU GPU comparision

+++ geometry/proximity/sycl/sycl_hydroelastic_surface.h
@@ -63,7 +63,7 @@ class SYCLHydroelasticSurface {
   }
 
   // Utility methods
-  size_t num_polygons() const { return area_.size(); }
+  uint32_t num_polygons() const { return area_.size(); }
   bool empty() const { return area_.empty(); }
 
  private:
